admin: colors via theme; refactor Relay1 and Relay2 into their own files
